+++ src/app/(pages)/@modal/(.)auth/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from '@/features/auth/login/ui/LoginForm';
+import { LoginForm } from '@/features/auth/login/ui/loginForm';
 
 export default function LoginPage() {
   console.log('modal입니다');
