+++ src/app/(pages)/@modal/(.)auth/signup/page.tsx
@@ -1,4 +1,4 @@
-import { SignupForm } from '@/features/auth/signup/ui/SignupForm';
+import { SignupForm } from '@/features/auth/signup/ui/signupForm';
 
 export default function SignUpPage() {
   return <SignupForm />;
