+++ src/app/(pages)/auth/login/page.tsx
@@ -1,6 +1,5 @@
-import { LoginForm } from '@/features/auth/login/ui/LoginForm';
+import { LoginForm } from '@/features/auth/login/ui/loginForm';
 
 export default function LoginPage() {
-  console.log('modal아닙닌다');
   return <LoginForm />;
 }
